breaking out tasks to handle Gitlab's inability to generate private tokens without manual interaction
https://gitlab.com/gitlab-org/gitlab-ce/issues/27954

If ran interactively, the role will prompt for a private token after a
user has been created. This can be generated manually using the web UI
and copied to the console prompt.

If ran non-interactively, the role will create the user and will not run
subsequent tasks that use the Gitlab API. The token can then be
generated, as above, before adding the token to the ansible run, i.e.

```
gitlab_private_token: 'DYDhs6_xgkW9r2F8YGpV'

```

The next run of the playbook will generate any defined projects and use ssh
keys.